일본어 지원 점수/페이즈 밸런스 조절

+++ Assets/01_Scripts/20_InGame/Scores/TimeManager.cs
@@ -19,7 +19,7 @@ public class TimeManager : MonoBehaviour {
 	private float starScale;
 	private bool progressChanging = false;
 
-	public int progressPerSecond = 10;
+	public float progressPerSecond = 10;
 	public float progressPerCube = 0.5f;
 	public float progressChangeSpeed = 5f;
